updated permission check
