+++ app/src/main/java/com/example/opensourceprj/ScanAdvertisementActivity.java
@@ -105,30 +105,15 @@ public class ScanAdvertisementActivity extends AppCompatActivity {
     private static final List<String> raspberryPiAddrList_4 = new ArrayList<>(Arrays.asList(raspberryPiAddr_4));
     private static final List<String> raspberryPiAddrList_5 = new ArrayList<>(Arrays.asList(raspberryPiAddr_5));
     private static final List<String> raspberryPiAddrList_ta = new ArrayList<>(Arrays.asList(raspberryPiAddr_ta));
-    private static final int PERMISSION_REQUEST_CODE = 1000;
-    private static final int REQUEST_ENABLE_BLUETOOTH = 1;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_scan_advertisement);
 
-        checkAndRequestPermissions(); // 권한 확인
-
         // 객체 생성
         blead = BluetoothAdapter.getDefaultAdapter();
         toast = Toast.makeText(ScanAdvertisementActivity.this, null, Toast.LENGTH_SHORT);
-
-        if (blead == null) {
-            toast.setText("Bluetooth is not available");
-            toast.show();
-        }
-
-        // 블루투스 기능 비활성화 시 기능 활성화
-        if (!blead.isEnabled()) {
-            Intent enableBluetoothIntent = new Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE);
-            startActivityForResult(enableBluetoothIntent, REQUEST_ENABLE_BLUETOOTH);
-        }
 
         try { // 애플리케이션 시작 시 파일을 읽어 TextView 설정
             File file = new File(Environment.getExternalStorageDirectory().getAbsoluteFile() + "/store_test.csv");
@@ -630,31 +615,6 @@ public class ScanAdvertisementActivity extends AppCompatActivity {
         else if(raspberryPiAddrList_ta.contains(raspPiAddr)) sensorTeam = "ta";
 
         return sensorTeam;
-    }
-
-    private void checkAndRequestPermissions() {
-        // 필요 권한 확인
-        String[] permissions = {
-                Manifest.permission.ACCESS_COARSE_LOCATION,
-                Manifest.permission.ACCESS_FINE_LOCATION,
-                Manifest.permission.BLUETOOTH_SCAN,
-                Manifest.permission.BLUETOOTH_CONNECT,
-                Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                Manifest.permission.READ_EXTERNAL_STORAGE,
-                Manifest.permission.MANAGE_EXTERNAL_STORAGE
-        };
-
-        List<String> permissionsNeeded = new ArrayList<>();
-        for (String permission : permissions) {
-            if (ContextCompat.checkSelfPermission(this, permission) != PackageManager.PERMISSION_GRANTED) {
-                permissionsNeeded.add(permission);
-            }
-        }
-
-        if (!permissionsNeeded.isEmpty()) {
-            // 사용자에게 권한 요청 다이얼로그 표시
-            ActivityCompat.requestPermissions(this, permissionsNeeded.toArray(new String[0]), PERMISSION_REQUEST_CODE);
-        }
     }
 
     private class NetworkTask extends AsyncTask<Void, Void, Document> {
